% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidirection_mvmr.R
\name{bidirection_mvmr}
\alias{bidirection_mvmr}
\title{Conduct bidirection MVMR between traits}
\usage{
bidirection_mvmr(
  ex_dat1,
  ex_dat2,
  ex_dat3,
  ex_dat4,
  min_instruments = 3,
  effect_size_cutoff = 0.1,
  R2_cutoff = 0.85,
  df_info = NULL,
  type_list = c("continuous", "continuous"),
  prevalence_list = NULL
)
}
\arguments{
\item{ex_dat1}{Dataframe of instruments for the main exposure or outcome and extra trait (X/Y + M). Output from TwoSampleMR::mv_extract_exposures()}

\item{ex_dat2}{Dataframe of instruments for a candidate confounder trait (Z). Output from TwoSampleMR::extract_instruments()}

\item{ex_dat3}{Dataframe of instruments for a candidate confounder trait and extra trait (Z + M). Output from TwoSampleMR::mv_extract_exposures()}

\item{ex_dat4}{Dataframe of instruments for the main exposure or outcome (X/Y). Output from TwoSampleMR::extract_instruments()}

\item{min_instruments}{minimum number of instruments for candidate traits. Default = 3}

\item{effect_size_cutoff}{Standardized effect size threshold. Default = 0.1}

\item{R2_cutoff}{R2 cutoff for duplicated traits with X or Y. Default = 0.85}

\item{df_info}{Dataframe of trait info containing sample sizes. The required columns include
`id` for trait ID, `sample_size` for sample sizes. Default = NULL}

\item{type_list}{A vector for the type of traits (X/Y + M). The order should be exactly matched
with traits. eg. c("binary","continuous") for the first exposure is a binary trait}

\item{prevalence_list}{A vector for prevalence of traits (X/Y + M). The order should
be exactly matched with exposures. For continuous trait, just write NA. eg. c(0.1, NA)}
}
\value{
A list contain bidirection estimates and traits correlation
}
\description{
Conduct bidirection MVMR between traits
}
