% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_Lasso.R
\name{corrected_Lasso}
\alias{corrected_Lasso}
\title{Use stepwise selection to do confounder selection}
\usage{
corrected_Lasso(
  id_exposure,
  id.list,
  df_info,
  mvdat_y,
  radius_type = "1se",
  seed = 1,
  maxits = 1e+06
)
}
\arguments{
\item{id_exposure}{GWAS ID of the main exposure}

\item{id.list}{GWAS ID list of traits based on previous steps}

\item{df_info}{Dataframe of trait info from previous steps}

\item{mvdat_y}{Harmonized data for the outcome produced by select_instruments step}

\item{radius_type}{radius type for corresponding loss. Default = "1se"}

\item{seed}{Default = 1}

\item{maxits}{maximum number of iterations for algorithm converges. Default=1000000}
}
\value{
A GWAS ID vector and a trait info dataframe
}
\description{
Use stepwise selection to do confounder selection
}
