% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_Lasso.R
\name{double_Lasso}
\alias{double_Lasso}
\title{Use classic Double Lasso to do confounder selection}
\usage{
double_Lasso(
  id_exposure,
  id.list,
  df_info,
  mvdat_x,
  mvdat_y,
  lambda_type = "min",
  seed = 1
)
}
\arguments{
\item{id_exposure}{GWAS ID of the main exposure}

\item{id.list}{GWAS ID list of traits based on previous steps}

\item{df_info}{Dataframe of trait info from previous steps}

\item{mvdat_x}{Harmonized data for the main exposure produced by select_instruments step}

\item{mvdat_y}{Harmonized data for the outcome produced by select_instruments step}

\item{lambda_type}{Lasso penalty type. Default = "min"}

\item{seed}{Default = 1}
}
\value{
A GWAS ID vector and a trait info dataframe
}
\description{
Use classic Double Lasso to do confounder selection
}
