% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_filter.R
\name{downstream_filter}
\alias{downstream_filter}
\title{Filter downstream traits}
\usage{
downstream_filter(
  id_exposure,
  id.list,
  df_info,
  res,
  p1 = 0.05,
  p2 = 0.01,
  MR_method
)
}
\arguments{
\item{id_exposure}{GWAS ID of the main exposure}

\item{id.list}{GWAS ID list of traits based on previous steps}

\item{df_info}{Dataframe of trait info from previous steps}

\item{res}{Bidirection MR results for selected traits}

\item{p1}{p-value cutoff for adjusted pval to select upstream traits. Default=0.05}

\item{p2}{p-value cutoff for adjusted pval to select downstream traits. Default=0.01}

\item{MR_method}{Bidirection MR or MVMR methods. Options: MR_IVW, MR_GRAPPLE,
MR_MRBEE, MVMR_IVW, MVMR_GRAPPLE, MVMR_MRBEE}
}
\value{
A GWAS ID vector, a trait info dataframe, a trait dataframe with four direction
estimate and t-test results
}
\description{
Use Bidirectional MR or MVMR results to filter out downstream traits.
It's based on FDR-adjusted pvalue on four directions. Select either upstream traits
for exposure or outcome direction (p_ZtoX_adj < p1 | p_ZtoY_adj < p1) and
exclude all downstream traits for both directions (p_XtoZ_adj < p2 | p_YtoZ_adj < p2).
}
