% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_pval.R
\name{estimate_R_pval}
\alias{estimate_R_pval}
\title{Estimate genetic correlation matrix by pvalue}
\usage{
estimate_R_pval(beta_files, p_thresh = 0.05)
}
\arguments{
\item{beta_files}{Paths of merged GWAS summary data after LD pruning}

\item{p_thresh}{pvalue cutoff for filtering data}
}
\value{
A matrix with pairwise sample overlapping correlation between traits
}
\description{
Estimate genetic correlation matrix by pvalue
}
