% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_combine.R
\name{format_combine}
\alias{format_combine}
\title{Combine and harmonize local GWAS summary data with multiple traits}
\usage{
format_combine(
  id_list,
  file_list,
  out_dir = NULL,
  prefix = NULL,
  snp = NA,
  pos = NA,
  chrom = NA,
  A1 = NA,
  A2 = NA,
  beta_hat = NA,
  se = NA,
  p_value = NA,
  af = NA,
  sample_size = NA,
  effect_is_or = "no"
)
}
\arguments{
\item{id_list}{GWAS ID list. The first should be the outcome ID.}

\item{file_list}{Directory path of formatted GWAS summary data. Default is in the current work directory.}

\item{out_dir}{Output data path. Default is in the current work directory.}

\item{prefix}{Name prefix for the output. Default = NULL}
}
\value{
Save one dataframe per chromosome with columns for SNP info
}
\description{
Combine and harmonize local GWAS summary data with multiple traits
}
