% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gwas_data.R
\name{format_gwas_data}
\alias{format_gwas_data}
\title{Format GWAS summary statistics from ieu-OpenGWAS database}
\usage{
format_gwas_data(data_file, output_file)
}
\arguments{
\item{data_file}{Raw GWAS summary data path}

\item{output_file}{Output data path}
}
\value{
Formatted GWAS summary data with .vcf.bgz
}
\description{
Format GWAS summary statistics from ieu-OpenGWAS database
}
