% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ldsc_prune.R
\name{format_ldsc_prune}
\alias{format_ldsc_prune}
\title{Format and combine GWAS summary data, do LDSC for R matrix and LD-pruning}
\usage{
format_ldsc_prune(
  df_file,
  df_info,
  r2_thresh,
  clump_kb,
  type,
  pthresh,
  ref_path,
  ld_files,
  m_files
)
}
\arguments{
\item{df_file}{A dataframe contain "id" (trait id) and "location" (GWAS summary data location) columns.}

\item{df_info}{Dataframe of trait info from previous step}

\item{r2_thresh}{Clumping r2 cut off. Default = 0.001}

\item{clump_kb}{Clumping distance cut off. Default = 10000}

\item{type}{LD clumping prioritization. Either pvalue or rank. Default = "pvalue"}

\item{pthresh}{pvalue threshold. Default = 1}

\item{ref_path}{Path for the LD reference panel.}

\item{ld_files}{Paths of reference LD score files}

\item{m_files}{Paths of reference M files}
}
\value{
A list of selected traits, a dataframe of conditional instrument strength and a dataframe of trait info
}
\description{
This is a combined function for a series of step: 1. Format and combine
GWAS summary statistics; 2. Do LD pruning for every chromosome and combine them
together; 3. Use LDSC to calculate sample overlap and genetic correlation matrix.
We recommend you to use this only when the trait list is short.
}
