% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_cluster.R
\name{greedy_cluster}
\alias{greedy_cluster}
\title{Greedy algorithm to get clusters.}
\usage{
greedy_cluster(id.list, R, df_info, R2_cutoff = 0.9)
}
\arguments{
\item{id.list}{GWAS ID list of traits based on previous steps}

\item{R}{Dataframe of pairwise correlation matrix. The colnames and rownames are trait IDs.}

\item{df_info}{Dataframe of trait ID info. It requires to contain `id`: trait ID,
`n_inst`: the number of instruments.}

\item{R2_cutoff}{high correlation cutoff to assign as duplicated traits. Default=.9}
}
\description{
Hierarchically clusters traits using a greedy algorithm.
Initially, traits are clustered according to the highest pairwise correlations
above a cutoff threshold. Then, within each cluster, traits are subclustered
based on descending order of instrument numbers with high correlations grouped together.
}
