% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_group.R
\name{greedy_group}
\alias{greedy_group}
\title{Greedy algorithm to get unique traits.}
\usage{
greedy_group(id.list, R, df_info, R2_cutoff = 0.9)
}
\arguments{
\item{id.list}{GWAS ID list of traits based on previous steps}

\item{R}{Dataframe of pairwise correlation matrix. The colnames and rownames are trait IDs.}

\item{df_info}{Dataframe of trait ID info. It requires to contain `id`: trait ID,
`n_inst`: the number of instruments.}

\item{R2_cutoff}{high correlation cutoff to assign as duplicated traits. Default=.9}
}
\description{
Hierarchically clusters traits using a greedy algorithm.
First rank traits according to the number of genome-wide significant variants after LD clumping. 
We then iteratively select the top trait and remove all traits with a similarity score greater than 
the threshold with the selected trait, continuing until all traits are either selected or removed.
}
