% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_prune_chrom_plink.R
\name{ld_prune_chrom_plink}
\alias{ld_prune_chrom_plink}
\title{LD clumping for harmonized data}
\usage{
ld_prune_chrom_plink(
  r2_thresh = 0.001,
  clump_kb = 10000,
  type = "pvalue",
  pthresh = 1,
  beta_dir = NULL,
  prefix = NULL,
  ref_path,
  out_dir = NULL
)
}
\arguments{
\item{r2_thresh}{Clumping r2 cut off. Default = 0.001}

\item{clump_kb}{Clumping distance cut off. Default = 10000}

\item{type}{LD clumping prioritization. Either pvalue or rank. Default = "pvalue"}

\item{pthresh}{pvalue threshold. Default = 1}

\item{beta_dir}{Directory path of merged data. Default is the current work directory.}

\item{prefix}{Name prefix for the output. Default = NULL}

\item{ref_path}{Path for the LD reference panel.}

\item{out_dir}{Output data path. Default is in the current work directory.}
}
\value{
Save one dataframe per chromosome with columns for SNP info
}
\description{
LD clumping for harmonized data
}
