% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_prune_plink.R
\name{ld_prune_plink}
\alias{ld_prune_plink}
\title{LD clumping for harmonized data}
\usage{
ld_prune_plink(
  X,
  r2_thresh = 0.001,
  clump_kb = 10000,
  type = "pvalue",
  pthresh = 1,
  ref_path
)
}
\arguments{
\item{X}{Paths of merged GWAS summary data by chromosomes}

\item{r2_thresh}{Clumping r2 cut off. Default = 0.001}

\item{clump_kb}{Clumping distance cut off. Default = 10000}

\item{type}{LD clumping prioritization. Either pvalue or rank. Default = "pvalue"}

\item{pthresh}{pvalue threshold. Default = 1}

\item{ref_path}{Path for the LD reference panel.}
}
\value{
One dataframe per chromosome with columns for SNP info
}
\description{
LD clumping for harmonized data
}
