% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc_full.R
\name{ldsc_full}
\alias{ldsc_full}
\title{Estimate genetic correlation matrix by LDSC method}
\usage{
ldsc_full(beta_files, ld_files, m_files)
}
\arguments{
\item{beta_files}{Paths of combined GWAS data without LD pruning}

\item{ld_files}{Paths of reference LD score files}

\item{m_files}{Paths of reference M files}
}
\value{
A list contain sample overlap matrix (Re) and genetic correlation matrix (Rg)
}
\description{
Estimate genetic correlation matrix by LDSC method
}
