% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVMR_analysis.R
\name{MVMR_analysis}
\alias{MVMR_analysis}
\title{MVMR analysis}
\usage{
MVMR_analysis(
  id_exposure,
  id_outcome,
  id.list,
  df_info,
  r2 = 0.001,
  kb = 10000,
  pval_threshold = 5e-08,
  find_proxies = TRUE,
  pop = "EUR",
  harmonise_strictness = 2,
  MVMR_method,
  pleio_p_thresh = 0,
  Rcor = NULL
)
}
\arguments{
\item{id_exposure}{GWAS ID of the main exposure}

\item{id_outcome}{GWAS ID of the outcome}

\item{id.list}{GWAS ID list of traits based on previous steps}

\item{df_info}{Dataframe of trait info from previous steps}

\item{r2}{LD-clump r2. Default = 0.001}

\item{kb}{LD-clump kb. Default = 10000}

\item{pval_threshold}{Instrument selection cutoff. Default = 5e-8}

\item{find_proxies}{Whether look for proxies. Default = TRUE}

\item{pop}{Super population to use. Default = "EUR"}

\item{harmonise_strictness}{Data harmonise strictness. See documentation of TwoSample MR}

\item{MVMR_method}{IVW, IVW_instrument_specific, MRBEE, GRAPPLE, ESMR}

\item{pleio_p_thresh}{P-value threshold for determining if a specific IV has
sufficient evidence of horizontal pleiotropy to be removed from causal estimation. Default=0}

\item{Rcor}{Correlation matrix for the outcome and all exposures}
}
\value{
A dataframe with trait estimates for correponding methods
}
\description{
MVMR analysis
}
