% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVMR_MRBEE.R
\name{MVMR_MRBEE}
\alias{MVMR_MRBEE}
\title{Use MRBEE to do MVMR analysis by locally data}
\usage{
MVMR_MRBEE(beta_files, R_matrix, pval_threshold = 5e-08)
}
\arguments{
\item{beta_files}{Paths of merged GWAS summary data after LD pruning}

\item{R_matrix}{Pairwise sample overlap matrix among traits}

\item{pval_threshold}{pvalue cutoff for selecting instruments. Default = 5e-8}
}
\value{
A dataframe of result summary
}
\description{
Use MRBEE to do MVMR analysis by locally data
}
