% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{quality_control}
\alias{quality_control}
\title{Quality control based on metadata}
\usage{
quality_control(
  id_exposure,
  dat,
  nsnp_cutoff = 1e+06,
  pop = "European",
  sex = "Males and Females",
  R2_cutoff = 0.9
)
}
\arguments{
\item{id_exposure}{GWAS ID for main exposure.}

\item{dat}{metadata for traits, a dataframe contains information of population, nsnp and etc.
If users input metadata by themselves, use make_metadata() to convert data format}

\item{nsnp_cutoff}{threshold of minimum number of SNPs. Default = 1e6}

\item{pop}{Limit population of traits. Default = "European"}

\item{sex}{Gender limitation. Default = "Males and Females"}

\item{R2_cutoff}{cutoff for duplicated traits of the exposure. Default = 0.9}
}
\value{
A GWAS ID vector and a trait info dataframe
}
\description{
Quality control based on metadata
}
