% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_instruments_local.R
\name{select_instruments_local}
\alias{select_instruments_local}
\title{Select instruments from local GWAS summary data for MVMR analysis}
\usage{
select_instruments_local(beta_files, pval_threshold = 5e-08)
}
\arguments{
\item{beta_files}{Paths of combined local GWAS data after LD clumping}

\item{pval_threshold}{Instrument selection cutoff. Default = 5e-8}
}
\value{
A list of harmonized data for the outcome (mvdat_y) and a list of harmonized data for the main exposure (mvdat_x)
}
\description{
Select instruments from local GWAS summary data for MVMR analysis
}
