% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strength_filter.R
\name{strength_filter}
\alias{strength_filter}
\title{Calculates the conditional F-statistic to assess instrument strength and filter traits}
\usage{
strength_filter(
  dat,
  dat_type = "local",
  R_matrix = NULL,
  df_info,
  pval_threshold = 5e-08,
  F_threshold = 5,
  Filter = FALSE,
  extra_traits = "None"
)
}
\arguments{
\item{dat}{A data frame of combined GWAS summary data after LD pruning.
For local data, the required columns include `SNP` for rsID, `trait_ID.beta` for beta hats,
`trait_ID.se` for standard errors, `trait_ID.p` for pvalues.
For IEU data, it should be output from TwoSampleMR::mv_harmonise_data().}

\item{dat_type}{Either "local" or "IEU". Default  = "local"}

\item{R_matrix}{Pairwise sample overlap matrix among traits. Default = NULL}

\item{df_info}{Dataframe of trait info from previous step}

\item{pval_threshold}{pvalue cutoff for selecting instruments. Default = 5e-8}

\item{F_threshold}{F-statistic cutoff. Default = 5}

\item{Filter}{Whether perform trait filtering based on F-stats and F_threshold. Default = FALSE}

\item{extra_traits}{trait ID you want to include no matter the instrument strength. Default = "None"}
}
\value{
A list of selected traits, a dataframe of conditional instrument strength and a dataframe of trait info
}
\description{
Calculates the conditional F-statistic to assess instrument strength and filter traits
}
