% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_filter_group.R
\name{string_filter_group}
\alias{string_filter_group}
\title{Use string similarity to filter traits}
\usage{
string_filter_group(
  id.list,
  df_info,
  df_inst_counts,
  R2_cutoff = 0.9,
  extra_traits = "None"
)
}
\arguments{
\item{id.list}{GWAS ID list of traits based on previous steps}

\item{df_info}{Dataframe of trait info from previous steps}

\item{df_inst_counts}{Dataframe of the number of instruments for each trait.
The required columns include `id` for trait ID, `n_inst` for the number of instruments.}

\item{R2_cutoff}{high correlation cutoff to assign as duplicated traits. Default=.9}

\item{extra_traits}{trait ID which you must want to include and select. Default = "None"}
}
\value{
A GWAS ID vector and a trait info dataframe
}
\description{
Keep traits with the highest number of instruments in one cluster. The trait correlation matrix is based on string similarity score and then use greedy cluster.
}
