% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_sim.R
\name{string_sim}
\alias{string_sim}
\title{Calculate trait similarity by trait name}
\usage{
string_sim(id.list, df_info)
}
\arguments{
\item{id.list}{GWAS ID list of traits based on previous steps}

\item{df_info}{Dataframe of trait info from previous steps}
}
\value{
df_pair: A dataframe contain string correlation for each pair. R_matrix: Pairwise correlation matrix
}
\description{
Calculate trait similarity by trait name
}
