% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_traits_inst.R
\name{unique_traits_inst}
\alias{unique_traits_inst}
\title{Select unique traits based on pairwise genetic correlation}
\usage{
unique_traits_inst(
  id.list,
  df_info,
  R_matrix,
  R2_cutoff = 0.8,
  extra_traits = "None"
)
}
\arguments{
\item{id.list}{GWAS ID list of traits based on previous steps}

\item{df_info}{Dataframe of trait info from previous steps}

\item{R_matrix}{Pairwise genetic correlation matrix. Colnames are trait names}

\item{R2_cutoff}{high correlation cutoff to assign as duplicated traits. Default=0.8}

\item{extra_traits}{trait ID which is adjusted for in bidirection MR. Default = "None"}
}
\value{
A GWAS ID vector and a trait info dataframe
}
\description{
Select unique traits based on pairwise genetic correlation
}
