% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidirection_mr.R
\name{bidirection_mr}
\alias{bidirection_mr}
\title{Conduct bidirection MR between traits}
\usage{
bidirection_mr(
  ex_dat1,
  ex_dat2,
  min_instruments = 3,
  effect_size_cutoff = 0.1,
  R2_cutoff = 0.85,
  type_list = c("continuous", "continuous"),
  prevalence_list = NULL
)
}
\arguments{
\item{ex_dat1}{Dataframe of instruments for the main exposure or outcome. Output from TwoSampleMR::extract_instruments()}

\item{ex_dat2}{Dataframe of instruments for a candidate confounder trait. Output from TwoSampleMR::extract_instruments()}

\item{min_instruments}{minimum number of instruments for candidate traits. Default = 3}

\item{effect_size_cutoff}{Standardized effect size threshold. Default = 0.1}

\item{R2_cutoff}{R2 cutoff for duplicated traits with X or Y. Default = 0.85}

\item{type_list}{A vector for the type of traits. The order should be exactly matched
with `ID1`, `ID2`. ID1 is the exposure of ex_dat1 and ID2 is the exposure of ex_dat2.
eg. c("binary","continuous") means that ID1 is a binary trait.}

\item{prevalence_list}{A list for prevalence of traits. The order should
be exactly matched with `ID1`, `ID2`. eg. list(0.1, NULL) means the prevalence of
ID1 is 0.1. For continuous trait, just write NULL.}
}
\value{
A list contain bidirection estimates and traits correlation
}
\description{
Conduct bidirection MR between traits
}
