% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic_Lasso.R
\name{classic_Lasso}
\alias{classic_Lasso}
\title{Use classic Lasso to do confounder selection}
\usage{
classic_Lasso(
  id_exposure,
  id_outcome,
  id.list,
  df_info,
  r2 = 0.001,
  kb = 10000,
  pval_threshold = 5e-08,
  find_proxies = TRUE,
  pop = "EUR",
  harmonise_strictness = 2,
  lambda_type = "min",
  seed = 1
)
}
\arguments{
\item{id_exposure}{GWAS ID of the main exposure}

\item{id_outcome}{GWAS ID of the outcome}

\item{id.list}{GWAS ID list of traits based on previous steps}

\item{df_info}{Dataframe of trait info from previous steps}

\item{r2}{LD-clump r2. Default = 0.001}

\item{kb}{LD-clump kb. Default = 10000}

\item{pval_threshold}{Instrument selection cutoff. Default = 5e-8}

\item{find_proxies}{Whether look for proxies. Default = TRUE}

\item{pop}{Super population to use. Default = "EUR"}

\item{harmonise_strictness}{Data harmonise strictness. See documentation of TwoSample MR}

\item{lambda_type}{Lasso penalty type. Default = "min"}

\item{seed}{Default = 1}
}
\value{
A GWAS ID vector and a trait info dataframe
}
\description{
Use classic Lasso to do confounder selection
}
