% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_filter.R
\name{downstream_filter}
\alias{downstream_filter}
\title{Filter downstream traits}
\usage{
downstream_filter(
  id_exposure,
  id.list,
  df_info,
  res,
  sig_level = 0.05,
  MR_method
)
}
\arguments{
\item{id_exposure}{GWAS ID of the main exposure}

\item{id.list}{GWAS ID list of traits based on previous steps}

\item{df_info}{Dataframe of trait info from previous steps}

\item{res}{Bidirection MR results for selected traits}

\item{sig_level}{One-sided t-test significant level. Default=0.05}

\item{MR_method}{Bidirection MR or MVMR methods. Options: MR_IVW, MR_GRAPPLE,
MR_ESMR, MR_MRBEE, MVMR_IVW, MVMR_GRAPPLE, MVMR_ESMR, MVMR_MRBEE}
}
\value{
A GWAS ID vector, a trait info dataframe, a trait dataframe with four direction
estimate and t-test results
}
\description{
Filter downstream traits
}
