% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_traits.R
\name{extract_traits}
\alias{extract_traits}
\title{Initial extract and filtering}
\usage{
extract_traits(
  id_exposure,
  id_outcome,
  batch = c("ieu-a", "ieu-b", "ukb-b", "ebi-a"),
  pop = "EUR",
  pval_x = 5e-08,
  pval_z = 1e-05,
  r2 = 0.001,
  kb = 10000,
  access_token = ieugwasr::check_access_token(),
  min_snps = 5
)
}
\arguments{
\item{id_exposure}{GWAS ID for main exposure.}

\item{id_outcome}{GWAS ID for the outcome.}

\item{batch}{GWAS database sub-batches, a vector.}

\item{pop}{Super-population to use as reference panel. Default = "EUR". Options are "EUR","SAS","EAS","AFR","AMR"}

\item{pval_x}{p-value threshold to extract instruments of the main exposure, default=5e-8}

\item{pval_z}{p-value threshold to retreive traits, default=1e-5}

\item{r2}{clumping r2 threshold. Default=0.001}

\item{kb}{clumping kb window. Default=10000}

\item{access_token}{Google OAuth2 access token. Default=check_access_token()}

\item{min_snps}{the number of minimum shared SNPs with IV of X. Default=5}
}
\value{
A GWAS ID vector and a trait info dataframe
}
\description{
Initial extract and filtering
}
