% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_cluster.R
\name{greedy_cluster}
\alias{greedy_cluster}
\title{Greedy algorithem to get clusters}
\usage{
greedy_cluster(id.list, R, R2_cutoff = 0.9)
}
\arguments{
\item{id.list}{GWAS ID list of traits based on previous steps}

\item{R}{Dataframe of pairwise correlation matrix. The colnames and rownames are trait IDs.}

\item{R2_cutoff}{high correlation cutoff to assign as duplicated traits. Default=.9}
}
\description{
Greedy algorithem to get clusters
}
