% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc_full.R
\name{ldsc_full}
\alias{ldsc_full}
\title{Estimate genetic correlation matrix}
\usage{
ldsc_full(l2_dir, beta_dir = NULL, out_dir = NULL, prefix = NULL)
}
\arguments{
\item{beta_dir}{Directory path of merged data. Default is the current work directory.}

\item{out_dir}{Output data path. Default is in the current work directory.}

\item{prefix}{Name prefix for the output. Default = NULL}

\item{ref_path}{Path for the LD reference panel.}
}
\value{
Save one dataframe per chromosome with columns for SNP info
}
\description{
Estimate genetic correlation matrix
}
