% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc_full.R
\name{ldsc_full}
\alias{ldsc_full}
\title{Estimate genetic correlation matrix by LDSC method}
\usage{
ldsc_full(dat, ld_files, m_files)
}
\arguments{
\item{dat}{Combined GWAS data without LD pruning. You have two options:
1. One dataframe of combined data
2. A vector for paths of combined data in each chromosome. It should have 22 elements.
eg: c("beta.1.RDS","beta.2.RDS",....,"beta.22.RDS")}

\item{ld_files}{Paths of reference LD score files}

\item{m_files}{Paths of reference M files}
}
\value{
A list contain sample overlap matrix (Re) and genetic correlation matrix (Rg)
}
\description{
Estimate genetic correlation matrix by LDSC method
}
