% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.R
\name{marginal}
\alias{marginal}
\title{Use marginal selection to do confounder selection}
\usage{
marginal(id.list, df_info, df_bidirection, p_cutoff = 0.05)
}
\arguments{
\item{id.list}{GWAS ID list of traits based on previous steps}

\item{df_info}{Dataframe of trait info from previous steps}

\item{df_bidirection}{Dataframe of trait with four direction estimates. Result from downstream_filter}

\item{p_cutoff}{pvalue threshold. Default = 0.05}
}
\value{
A GWAS ID vector and a trait info dataframe
}
\description{
Use marginal selection to do confounder selection
}
