% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVMR_ESMR.R
\name{MVMR_ESMR}
\alias{MVMR_ESMR}
\title{Use ESMR to do MVMR analysis by locally data}
\usage{
MVMR_ESMR(dat, R_matrix, pval_threshold = 5e-08, effect_size_cutoff = 0.1)
}
\arguments{
\item{dat}{A data frame of combined GWAS summary data after LD pruning.
The required columns include `SNP` for rsID, `trait_ID.beta` for beta hats,
`trait_ID.se` for standard errors, `trait_ID.z` for z-values, `trait_ID.ss` for sample sizes.}

\item{R_matrix}{Pairwise sample overlap matrix among traits}

\item{pval_threshold}{pvalue cutoff for selecting instruments. Default = 5e-8}

\item{effect_size_cutoff}{Standardized effect size threshold. Default = 0.1}
}
\value{
A dataframe of result summary
}
\description{
Use ESMR to do MVMR analysis by locally data
}
