% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVMR_GRAPPLE.R
\name{MVMR_GRAPPLE}
\alias{MVMR_GRAPPLE}
\title{Use GRAPPLE to do MVMR analysis by locally data}
\usage{
MVMR_GRAPPLE(beta_files, R_matrix, pval_threshold = 1e-05)
}
\arguments{
\item{beta_files}{Paths of merged GWAS summary data after LD pruning}

\item{R_matrix}{Pairwise sample overlap matrix among traits}

\item{pval_threshold}{pvalue cutoff for selecting instruments. Default = 1e-5}
}
\value{
A dataframe of result summary
}
\description{
Use GRAPPLE to do MVMR analysis by locally data
}
