% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVMR_GRAPPLE.R
\name{MVMR_GRAPPLE}
\alias{MVMR_GRAPPLE}
\title{Use GRAPPLE to do MVMR analysis by locally data}
\usage{
MVMR_GRAPPLE(
  dat,
  R_matrix,
  pval_threshold = 1e-05,
  type,
  ss.exposure = NULL,
  ss.outcome = NULL
)
}
\arguments{
\item{dat}{If type == "local", it's a data frame of combined GWAS summary data after LD pruning.
The required columns include `SNP` for rsID, `trait_ID.beta` for beta hats,
`trait_ID.se` for standard errors, `trait_ID.z` for zvalues, `trait_ID.p` for pvalues,
`trait_ID.ss` for sample sizes.
If type == "IEU", it should be the output from TwoSampleMR::mv_harmonise_data().}

\item{R_matrix}{Pairwise sample overlap matrix among traits}

\item{pval_threshold}{pvalue cutoff for selecting instruments. Default = 1e-5}

\item{type}{Input data type. It could be either "local" for local GWAS summary data after LD pruning or
"IEU" for output from TwoSampleMR::mv_harmonise_data().}

\item{ss.exposure}{A vector of sample size for exposures. You can provide it when type = "IEU".
The order of it should be the same with beta hat matrix and se matrix. Default = NULL}

\item{ss.outcome}{A numeric number of sample size for the outcome. You can provide it when type = "IEU". Default = NULL}
}
\value{
A dataframe of result summary
}
\description{
Use GRAPPLE to do MVMR analysis by locally data
}
