% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVMR_IVW.R
\name{MVMR_IVW}
\alias{MVMR_IVW}
\title{Use IVW to do MVMR analysis by locally data}
\usage{
MVMR_IVW(beta_files, pval_threshold = 5e-08)
}
\arguments{
\item{beta_files}{Paths of merged GWAS summary data after LD pruning}

\item{pval_threshold}{pvalue cutoff for selecting instruments. Default = 5e-8}
}
\value{
A dataframe of result summary
}
\description{
Use IVW to do MVMR analysis by locally data
}
