% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVMR_IVW.R
\name{MVMR_IVW}
\alias{MVMR_IVW}
\title{Use IVW to do MVMR analysis by locally data}
\usage{
MVMR_IVW(
  dat,
  pval_threshold = 5e-08,
  type,
  ss.exposure = NULL,
  ss.outcome = NULL,
  effect_size_cutoff = 0.1,
  type_outcome = "continuous",
  prevalence_outcome = NULL,
  type_exposure = NULL,
  prevalence_exposure = NULL
)
}
\arguments{
\item{dat}{If type == "local", it's a data frame of combined GWAS summary data after LD pruning.
The required columns include `SNP` for rsID, `trait_ID.beta` for beta hats,
`trait_ID.se` for standard errors, `trait_ID.z` for zvalues, `trait_ID.p` for pvalues,
`trait_ID.ss` for sample sizes
If type == "IEU", it should be the output from TwoSampleMR::mv_harmonise_data().}

\item{pval_threshold}{pvalue cutoff for selecting instruments. Default = 5e-8}

\item{type}{Input data type. It could be either "local" for local GWAS summary data after LD pruning or
"IEU" for output from TwoSampleMR::mv_harmonise_data().}

\item{ss.exposure}{A vector of sample size for exposures. You can provide it when type = "IEU".
The order of it should be the same with beta hat matrix and se matrix. Default = NULL}

\item{ss.outcome}{A numeric number of sample size for the outcome. You can provide it when type = "IEU". Default = NULL}

\item{effect_size_cutoff}{Standardized effect size threshold. Default = 0.1}

\item{type_outcome}{It could be either "continuous" or "binary". Default = "continuous"}

\item{prevalence_outcome}{Outcome prevalence. It should be input if the outcome is a binary trait. Default = NULL}

\item{type_exposure}{A vector for the type of exposures. The order should be exactly matched
with exposures. eg. c("continuous","binary","continuous") for the second exposure is a binary trait}

\item{prevalence_exposure}{A vector for prevalence of exposures. The order should
be exactly matched with exposures. For continuous trait, just write NA. eg. c(NA, 0.1, NA)}
}
\value{
A dataframe of result summary
}
\description{
Use IVW to do MVMR analysis by locally data
}
