% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVMR_MRBEE.R
\name{MVMR_MRBEE}
\alias{MVMR_MRBEE}
\title{Use MRBEE to do MVMR analysis by locally data}
\usage{
MVMR_MRBEE(dat, R_matrix, pval_threshold = 5e-08, pleio_threshold = 0, type)
}
\arguments{
\item{dat}{If type == "local", it's a data frame of combined GWAS summary data after LD pruning.
The required columns include `SNP` for rsID, `trait_ID.beta` for beta hats, `trait_ID.se` for standard errors, `trait_ID.p` for pvalues.
If type == "IEU", it should be the output from TwoSampleMR::mv_harmonise_data().}

\item{pval_threshold}{pvalue cutoff for selecting instruments. Default = 5e-8}

\item{pleio_threshold}{pvalue threshold in pleiotropy detection. Default = 0}

\item{type}{Input data type. It could be either "local" for local GWAS summary data after LD pruning or
"IEU" for output from TwoSampleMR::mv_harmonise_data().}
}
\value{
A dataframe of result summary
}
\description{
Use MRBEE to do MVMR analysis by locally data
}
