% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise.R
\name{stepwise}
\alias{stepwise}
\title{Use stepwise selection to do confounder selection}
\usage{
stepwise(id_exposure, id.list, df_info, mvdat_y, method = "forward")
}
\arguments{
\item{id_exposure}{GWAS ID of the main exposure}

\item{id.list}{GWAS ID list of traits based on previous steps}

\item{df_info}{Dataframe of trait info from previous steps}

\item{mvdat_y}{Harmonized data for the outcome produced by select_instruments step}

\item{method}{either forward, backward or both. Default = "forward"}
}
\value{
A GWAS ID vector and a trait info dataframe
}
\description{
Use stepwise selection to do confounder selection
}
