% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise.R
\name{stepwise}
\alias{stepwise}
\title{Use stepwise selection to do confounder selection}
\usage{
stepwise(
  id_exposure,
  id_outcome,
  id.list,
  df_info,
  method = "forward",
  r2 = 0.001,
  kb = 10000,
  pval_threshold = 5e-08,
  find_proxies = TRUE,
  pop = "EUR",
  harmonise_strictness = 2
)
}
\arguments{
\item{id_exposure}{GWAS ID of the main exposure}

\item{id_outcome}{GWAS ID of the outcome}

\item{id.list}{GWAS ID list of traits based on previous steps}

\item{df_info}{Dataframe of trait info from previous steps}

\item{method}{either forward, backward or both. Default = forward}

\item{r2}{LD-clump r2. Default = 0.001}

\item{kb}{LD-clump kb. Default = 10000}

\item{pval_threshold}{Instrument selection cutoff. Default = 5e-8}

\item{find_proxies}{Whether look for proxies. Default = TRUE}

\item{pop}{Super population to use. Default = "EUR"}

\item{harmonise_strictness}{Data harmonise strictness. See documentation of TwoSample MR}
}
\value{
A GWAS ID vector and a trait info dataframe
}
\description{
Use stepwise selection to do confounder selection
}
