% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strength_filter.R
\name{strength_filter}
\alias{strength_filter}
\title{Calculates the conditional F-statistic to assess instrument strength and filter traits}
\usage{
strength_filter(
  beta_files,
  R_matrix,
  df_info,
  pval_threshold = 5e-08,
  F_threshold = 5,
  extra_traits = "None"
)
}
\arguments{
\item{beta_files}{Paths of merged GWAS summary data after LD pruning}

\item{R_matrix}{Pairwise sample overlap matrix among traits}

\item{df_info}{Dataframe of trait info from previous step}

\item{pval_threshold}{pvalue cutoff for selecting instruments. Default = 5e-8}

\item{F_threshold}{F-statistic cutoff. Default = 5}

\item{extra_traits}{trait ID you want to include no matter the instrument strength. Default = "None"}
}
\value{
A list of selected traits, a dataframe of conditional instrument strength and a dataframe of trait info
}
\description{
Calculates the conditional F-statistic to assess instrument strength and filter traits
}
