% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_traits.R
\name{unique_traits}
\alias{unique_traits}
\title{Calculate pairwise correlation and select unique traits}
\usage{
unique_traits(
  id.list,
  df_info,
  cor_cutoff = 0.9,
  method,
  cluster_method = "complete",
  ncluster = NULL,
  cluster_plot = FALSE
)
}
\arguments{
\item{id.list}{GWAS ID list of traits based on previous steps}

\item{df_info}{Dataframe of trait info from previous steps}

\item{cor_cutoff}{high correlation cutoff to assign as duplicated traits. Default=.9}

\item{method}{filtering duplicate method: sample_size, nsnp, cluster}

\item{cluster_method}{the same with hclust method}

\item{ncluster}{The number of clusters. If not input by users, we'll use gap stats to
determine the optimal number of clusters}

\item{cluster_plot}{whether plot or not}
}
\value{
A GWAS ID vector and a trait info dataframe
}
\description{
Calculate pairwise correlation and select unique traits
}
