% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_traits.R
\name{unique_traits}
\alias{unique_traits}
\title{Calculate pairwise correlation and select unique traits}
\usage{
unique_traits(
  id.list,
  df_info,
  R_matrix,
  df_pairs,
  R2_cutoff = 0.8,
  method = "cluster",
  cluster_selection_method = "n_inst",
  extra_traits = "None",
  df_bidirection = NULL
)
}
\arguments{
\item{id.list}{GWAS ID list of traits based on previous steps}

\item{df_info}{Dataframe of trait info from previous steps}

\item{R_matrix}{Pairwise genetic correlation matrix. Colnames are trait names}

\item{df_pairs}{A dataframe contain string correlation for each pair}

\item{R2_cutoff}{high correlation cutoff to assign as duplicated traits. Default=0.8}

\item{method}{Filtering duplicate method: sample_size, nsnp, cluster. Default = "cluster"}

\item{cluster_selection_method}{Trait selection method in each cluster: n_inst (select traits with
the largest number of instruments), pvalue (-log10(p) mean for X and Y). Default = "n_inst"}

\item{extra_traits}{trait ID which is adjusted for in bidirection MR. Default = "None"}

\item{df_bidirection}{Dataframe of bidirection result. It should be input if you choose pvalue method for cluster selection.}
}
\value{
A GWAS ID vector and a trait info dataframe
}
\description{
Calculate pairwise correlation and select unique traits
}
